main/backend/requirements.txt
# HTTP and API libraries
fastapi==0.115.12
uvicorn==0.34.2
httpx>=0.24.0,<1.0.0
requests==2.32.3
pydantic==2.11.5

# Audio processing
librosa==0.11.0
soundfile==0.13.1
pydub==0.25.1
torchaudio>=2.0.0
torch>=2.0.0
numpy>=1.24.0
scipy==1.15.3

# Speech analysis and disfluency detection
webrtcvad==2.0.10
# pyaudioanalysis==0.3.14  # May need compilation, skip for now
# praat-parselmouth>=0.4.5  # May need compilation, skip for now

# Forced alignment for phoneme-level filler detection
forcealign>=1.1.9

# ML and AI
torch==2.7.1
transformers==4.52.4
# scikit-learn==1.3.2  # Already included as librosa dependency

# Speech recognition and analysis
vosk==0.3.44

# AI API client
openai==1.82.1

# Database
supabase==2.15.2

# Utilities
python-multipart==0.0.20
# python-jose[cryptography]==3.3.0  # May need compilation, skip for now
python-dotenv==1.1.0

# For future, consider adding a specific ASR/phoneme library here, e.g.:
# allosaurus
# pydub # if mp3 support is added and ffmpeg is available 